backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
httpx>=0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
backboard-sdk
livekit-api>=1.0.0
redis>=5.0.0
loguru>=0.7.0
python-jose[cryptography]>=3.3.0
